style: Enhance UserCard component styling for improved visual appeal and user interaction
